Got some instant instruction stuff going!
Needs lots of work on cleaning up each instruction struct and removing unnecessary stuff in them... and also cleaning up motor and heater instructions. Need to update the motor and heater controls and how to enable and check those instructions... but the user control and instant instruction processing is in!