Reads Program
Fixed homing and some calculations from the program. Moves the right number! Need to test speed and need to update movements for extruders and aux (only did axis motors), but should be easy. Need to finish saving and loading and sending config file and getting it right with configuring new home