Got config mainly going!
Still need to finish getting z value... It's not a valid number right now. Need to decide how I want to save and reference the z axes and sizes... cause they can be different. And should they only reference off the top? or stay accurate to top and bottom? Should the home point by zero? I think it should be... not how it is now.... But config state machine works in this! Weird lockup condition can occur sometimes...